+++ DICTIONARIES/Small Dictionary.txt
@@ -1,0 +1,4 @@
+wait
+waiter
+shop
+shopper
